Moved price processing methods nearer to usage + fixed a potential bug with markups applying
